Restructure backend module
