--- office-backend/src/main/kotlin/flock/community/office/monitoring/backend/controller/Root.kt
@@ -1,20 +1,0 @@
-package flock.community.office.monitoring.backend.controller
-
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/api")
-class Root {
-
-    @GetMapping
-    fun getGreeting(@RequestParam name: String?) = Greeting(name)
-}
-
-data class Greeting constructor(
-    private val name: String?,
-) {
-    val greeting: String get() = if (name.isNullOrBlank()) "H🤦‍♀️llo  World!" else "Hello $name!"
-}
