+++ office-backend/src/main/kotlin/flock/community/office/monitoring/backend/device/repository/DeviceStateRepository.kt
@@ -1,0 +1,15 @@
+package flock.community.office.monitoring.backend.device.repository
+
+import org.springframework.cloud.gcp.data.datastore.repository.DatastoreRepository
+import org.springframework.stereotype.Repository
+import java.time.Instant
+
+@Repository
+interface DeviceStateRepository : DatastoreRepository<DeviceStateEntity, String>{
+    fun findTop100ByDeviceIdOrderByDateDesc(deviceId: String): Iterable<DeviceStateEntity>
+    fun findAllByDeviceIdAndDateGreaterThanOrderByDateAsc(deviceId: String, from: Instant): Iterable<DeviceStateEntity>
+}
+
+
+
+
